[#154064861] Updates appearance of page header
Makes it so that the page header will use a sans-serif title styling when breadcrumbs are present, and use a serif styling if there are no breadcrumbs. Also blends the page header background color into the current site background color. There will soon be site-wide layout and styling changes that will put in place rules for how we do the background color for the page header. For now, this color change helps the header to look more harmonious until those site-wide changes happen.