Features/page header integration #154064861 (#21)
* [#154064861] - Fixed dropdownmenuitem and dropdownmenuitemcheckbox naming

* [#154064861] - Removed wrapper for lease ups header

* [#154064861] Adds latest PL CSS; slightly advances page header integration

* [#154064861] Adds a layout partial for page header

* [#154064861] - Update jest tests

* [#154064861] using PageHeader with React in Listings

* [#154064861] Applies new structure to pages in React

* [#154064861] Updates appearance of page header

Makes it so that the page header will use a sans-serif title styling when breadcrumbs are present, and use a serif styling if there are no breadcrumbs. Also blends the page header background color into the current site background color. There will soon be site-wide layout and styling changes that will put in place rules for how we do the background color for the page header. For now, this color change helps the header to look more harmonious until those site-wide changes happen.

* [#154064861] Updates tests; updates page header content prop types

* [#154064861] Updates some applications-related pages

Updates the pages related to flagged applications to work with the new route structure that was just merged in from master. Also removes the now-unneeded application spreadsheet route/action/view.