[#155196338] Updates several aspects of Confirmed Preferences display
This commit: 1) Adds the ability for the ExpandableTable component to have classes applied to its cells. 2) Right-aligns the Preference Rank column content in the Confirmed Preferences table. 3) Makes it so that if a preference in the Confirmed Preferences table has Invalid status, it gets a red "x" icon; if it has Confirmed status, it gets a green check icon; and if it has any status other than Invalid or Confirmed, it gets no icon. 4) Also does some minor formatting cleanup