[#157785070] Adds fallback for DOB field for browsers that don't support date input type
We are using the HTML5 "date" input field type on the DOB fields. However, not all browsers support this input type. Notably, Safari and IE do not support it. DAHLIA supports IE10+ and Safari. So this commit adds attributes to the DOB fields to ensure that the fields produce the same values in browsers that do not support the "date" input type. Please see this MDN doc for more info: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/date.