Features/update date input #157785070 (#60)
* [#157785070] - Updated date input for primary applicant and household member

- Removed use of datepicker
- Created error messages/validation on field touch and submit
- Scrolls user up to the stop when form is invalid on submit

* [#157785070] - Remove datepicker usage sitewide

* [#157785070] - Updated tests

* [#157785070] Adds fallback for DOB field for browsers that don't support date input type

We are using the HTML5 "date" input field type on the DOB fields. However, not all browsers support this input type. Notably, Safari and IE do not support it. DAHLIA supports IE10+ and Safari. So this commit adds attributes to the DOB fields to ensure that the fields produce the same values in browsers that do not support the "date" input type. Please see this MDN doc for more info: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/date.

* [#157785070] Add inline validation for valid date

* [#157785070] - Remove console.log

* [#157785070] Add better date input validation