[#158503202] Uses new custom API endpoint for deleting rental assistances
Using the standard REST API, it is not possible for a user with lower permissions to delete a rental assistance that was created by a user with higher permissions. The custom API endpoint provided by Vertiba allows a user with lower permissions to delete a rental assistance even if it was added by a user with higher permissions.