[#161196317] Updates unit test for annual income field validation
We had been using a snapshot of the entire PaperApplicationForm to check that the annual income field received the correct error validation. This commit does away with that, saving us a few thousand lines of snapshot, and instead checks the PaperApplicationForm component wrapper directly for the presence of the annual income field validation states.