Bugs/commas in number fields #161196317 (#158)
* SAVEPOINT

* [#161196317] Validating and parsing annual income

* [#161196317] Lint fixed

* [#161196317] Do not validate anual income if blank

* [#161196317] Convert to float

* [#161196317] Updates regex for annual income to be a whitelist

Also updates related unit test.

* [#161196317] Updates paper app annual income field error msg

* [#161196317] Adds unit test for validation of annual income field on paper app

* [#161196317] Update snapshots after merge from master

* [#161196317] Updates unit test for annual income field validation

We had been using a snapshot of the entire PaperApplicationForm to check that the annual income field received the correct error validation. This commit does away with that, saving us a few thousand lines of snapshot, and instead checks the PaperApplicationForm component wrapper directly for the presence of the annual income field validation states.

* [#161196317] Updates attrs on modals in snapshots

This reverses a previous update to the snapshotted modals because that update was found not to work in all cases.