Bugs/rental assistance delete error #162430991 #166
[#162430991] Disables rental assistance deletion

The rental assistance delete functionality is only possible via a custom Salesforce API endpoint that allows rental assistance deletion. This Salesforce work is not yet deployed to Prod, so we are disabling the rental assistance delete function by hiding the button. When the Salesforce work is on Prod, we will re-enable rental assistance deletion. This PR also contains the disabling or removal of any tests or parts of tests that involved rental assistance delete functionality.