[162457306] WIP - refactor labelize function for clarity
Also standardize naming of index variables in the preference form, and remove an unused variable from a function in the preference form. This commit marked WIP because I was unable to test it due to Salesforce being down. Pushing for other team members to review in the meantime. Will test when Salesforce is back up.