Fix AMI percent input bug #172688620
This change makes the supplemental application form submit properly when
the AMI percent input text has a percent sign in it. Additionally, this
change makes the behavior of the percent input closer to that of the
currency input, where a percent sign is automatically added after a
user defocuses the view, and the input only allows numbers (no decimal
points) so a failure case ends up being pretty much impossible.