Fix submission type #171270123 (#354)
* Fix submission type #171270123

* Fix "Any Listing" + "Any Type" options #172804251

We were sending "Any Listing" and "Any Type" as selection values to the
backend. This is not the expected behavior, as those are labels, not
values, and their values should be null. The issue was the way the form
fields render with null values, they pass the label as the value on
submit. This change makes it so that we always pass null when the value
is null.

Additionally, we add filtering to scrub requests with empty strings as
values, because on this page all empty strings mean "don't filter".

* Fix select fields for lease up application filter

* update snapshot

Co-authored-by: Andrea Egan <akegan@users.noreply.github.com>
Co-authored-by: James Wills <james.wills@sfgov.org>