Separate AMI chart and year dropdowns #172864201
Previously we combined AMI charts and years, but recently we found that
agents want to use more recent years than our options provided. This
change makes it so that agents will always be able to choose the current
year, even if it's not included in our AMI chart data. Additionally we
split up the year and AMI chart dropdowns because there is no longer a
1:1 mapping between the two.